В базе беньги хранятся в decimal, в яве в double. В базе переделана таблица заявок на оплату. в таблицу добавлено больше статистики о произведенной оплате

+++ trunk/pokerweb/src/java/com/pokerweb/Area/FieldOutMoney.java
@@ -11,8 +11,8 @@ package com.pokerweb.Area;
 public class FieldOutMoney {
     public int Id;
     public String Login;
-    public Float Sum;
+    public double Sum;
     public String Date;
-    public float Balance;
+    public double Balance;
     public boolean Processed;
 }
